+++ src/schedule/schedule.h
@@ -9,7 +9,7 @@
 void sched_init_proc(process_t *proc);
 void sched_set_core(process_t *proc);
 void sched_set_runable(process_t *proc);
-void sched_set_wait(process_t *proc, size_t pid);
+void sched_set_wait(process_t *proc, int pid);
 void sched_set_exit(process_t *proc);
 
 void schedule(int core_id);
