+++ src/syscall/syscalls/sys_wait.cpp
@@ -6,7 +6,7 @@
 #include "sys_yield.h"
 #include "../../process/process.h"
 
-sys_wait::sys_wait(size_t pid)
+sys_wait::sys_wait(int pid)
 	:syscall_t(syscall_id_t::WAIT)
 {
 	this->pid = pid;
